Move saveImage function to Utils.kt
